requirements.txt
langchain==0.3.21
langchain-community==0.3.20
streamlit==1.43.2
psycopg2-binary==2.9.10
chromadb==0.6.3
faiss-cpu==1.10.0
sentence-transformers==3.4.1
huggingface-hub==0.29.3
pydantic==2.10.6
python-dotenv==1.0.1
pypdf==5.4.0
docx2txt==0.9
rank-bm25==0.2.2
scikit-learn==1.6.1
gensim==4.3.3
fasttext==0.9.3